address review feedback from aubin-tchoi on PR #29135
- fix availability snapshot generator to emit the existing compact
  one-entry-per-line format instead of expanding every object, so the
  diff only shows the actual servicenow addition
- extract magic default incident limit into a named constant
- inline single-use endpoint variable
- drop log call redundant with withToolLogging
- categorize API errors by status code (5xx tracked, 4xx untracked)
- normalize the instance URL once, at the point it's read from auth
  metadata, instead of on every API call
- untrack the "please reconnect" config error
- rename helpers.ts to client.ts, keep only API-interaction code there,
  move renderIncident (pure formatting) to a new helpers.ts, and
  replace the withAuth wrapper with getServiceNowCredentials to match
  the pattern already used by the freshservice server
